feat(api): introduces the batch_job endpoint and all the miscellaneous infra & tests around it (#6)
PR setups batch_job endpoint, allowing users to request tx_fees for some past period of time. The next PR will feature the job executor. The PR also introduces redis as a messaging queue — both in docker-compose and the CI (so that tests related to it can run properly)

The flow goes like this

- client submits a POST /batch_job with UNIX epoch start_time & end_time
- sanity checks executed on the input
- add an entry to the DB with the batch_job status as 'pending'
- send a message to our messaging queue (redis) that there's a new job that has to be processed
- [in the next PR we'll introduce] the job executor picks up the job and executes the historical fee search