feat(batch): add the historical fee batch executor (#7)
Adds the historical batch executor component, running concurrently with the live data fetcher & the REST API.

One of the future PRs will put the components behind flags such that we could run an instance without having to necessarily run all the components simultaneously. Esp. useful if u want to horizontally scale the REST API or the number of batch executors